- more bug fixes

+++ manager.js
@@ -28,12 +28,16 @@ export async function main(ns) {
 
     // ------------------ variables ------------------
     const updateInterval = 1000 * 60 * (getArgValue(ns.args, "u") || 20); // get update interval in minutes. Default is 20 minutes
-    const spendMoney = getArgValue(ns.args, "m") !== undefined ? getArgValue(ns.args, "m") : true; // do we spend money? Default is true
+    const spendMoney = getArgValue(ns.args, "m") !== null ? getArgValue(ns.args, "m") : true; // do we spend money? Default is true
     const freeRAMonHome = 50; // gb of RAM we want to keep free on home server for other scripts
     const serversFile = "data/servers.txt";
     const hackFile = "hack.js";
     const scanFile = "scan.js";
     const budgetFactorPurchaseServers = 0.9; // we spend at most 90% of our money on servers
+    const hackOnHome = false; // do we hack on the home server?
+
+    // ------------------ kill previous manager.js ------------------
+    killPreviousManagers(ns);
 
     // ------------------ check data and files ------------------
     // If we just started (i.e. our hacking level is 1), we remove the servers file to reset the log data
@@ -54,17 +58,20 @@ export async function main(ns) {
         // ------------------ update cycle ------------------
         scan(ns, scanFile); // scan all servers with scan.js to update the `servers.txt` file
         nukeServers(ns, serversFile); // nuke all servers
+        // todo: solve contracts where possible. 
         
-        // if spend money, purchase and upgrade servers
+        // if spendMoney = true, purchase and upgrade servers
         if (spendMoney) {
             scan(ns, scanFile);
             ns.run('purchase_server.js', 1, 'buy', purchaseServersBudget); // purchase/upgrade servers
         }
 
         // #####################################################
-        scan(ns, scanFile); // scan all servers with scan.js to update the `servers.txt` file
-        hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile); // hack on all purchased servers
+        scan(ns, scanFile);
         hackOnTargetServers(ns, serversFile); // hack on target servers
+        hackOnPurchasedServers(ns, freeRAMonHome, hackFile, serversFile, hackOnHome); // hack on all purchased servers
+
+        scan(ns, scanFile);
 
         await ns.sleep(updateInterval); // wait until the next update cycle
     }
@@ -81,6 +88,24 @@ function getArgValue(args, key) { // get the value of the argument key
     return null;
 }
 
+function killPreviousManagers(ns) { // kill all previous manager.js scripts
+    let processes = ns.ps('home');
+    let currentManagerPID;
+
+    // we looop through the processes on our home server from last to first. 
+    // the first process with "manager" in the filename is the current script, we leave that alone. 
+    // all other manager scripts will be killed
+    for (let i = processes.length - 1; i >= 0; i--) {
+        if (processes[i].filename.includes('manager') && !currentManagerPID) {
+            currentManagerPID = processes[i].pid;
+            continue;
+        }
+        if (processes[i].filename.includes('manager')) {
+            ns.kill(processes[i].pid);
+        }
+    }
+}
+
 function scan(ns, scanFile) { // scan all servers
     ns.run(scanFile, 1, 'scan');
 }
@@ -89,12 +114,22 @@ function nukeServers(ns, serversFile) { // nuke all servers
     const serverData = readData(ns, serversFile);
     const indexHostname = serverData[0].indexOf('hostname');
     const indexAdmin = serverData[0].indexOf('hasAdminRights');
+    const indexHackable = serverData[0].indexOf('hackable');
 
     for (let row = 1; row < serverData.length - 1; row++) {
-        if (!serverData[row][indexAdmin] && serverHackable(serverData[row][indexHostname], serverData)) { // if the server is hackable and we don't have admin rights yet
+        if (!serverData[row][indexAdmin] && serverData[row][indexHackable]) { // if the server is hackable and we don't have admin rights yet
+            openPorts(ns, serverData[row][indexHostname]); // we open the ports
             ns.nuke(serverData[row][indexHostname]); // we nuke it
         }
     }
+}
+
+function openPorts(ns, target) { // open all ports on the target server
+    if (ns.fileExists('BruteSSH.exe', 'home')) ns.brutessh(target);
+    if (ns.fileExists('FTPCrack.exe', 'home')) ns.ftpcrack(target);
+    if (ns.fileExists('relaySMTP.exe', 'home')) ns.relaysmtp(target);
+    if (ns.fileExists('HTTPWorm.exe', 'home')) ns.httpworm(target);
+    if (ns.fileExists('SQLInject.exe', 'home')) ns.sqlinject(target);    
 }
 
 function hackOnTargetServers(ns, serversFile) { // hack on target servers
@@ -128,22 +163,23 @@ function hackOnTargetServers(ns, serversFile) { // hack on target servers
     }
 }
 
-function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack on all purchased servers
+function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, serversFile, hackOnHome) { // hack on all purchased servers
     let script_ram = ns.getScriptRam(hackFile);
-    let serverData = readData(ns, scanFile);
+    let serverData = readData(ns, serversFile);
     let indexPurchasedByPlayer = serverData[0].indexOf('purchasedByPlayer');
     let indexMaxRam = serverData[0].indexOf('maxRam');
     let indexServerValue = serverData[0].indexOf('serverValue');
     let indexHackable = serverData[0].indexOf('hackable');
     let indexHostname = serverData[0].indexOf('hostname');
     let totalRunnableThreads = Math.floor((ns.getServerMaxRam('home') - freeRAMonHome) / script_ram); // calculate available RAM on home server
-
+    
     // ------------------ calculate total runnable threads ------------------
+    if (!hackOnHome) totalRunnableThreads = 0; // if we don't want to hack on the home server, we set the threads to 0 instead
     if (totalRunnableThreads < 0) { // just in case our home server has less than what we defined as min free RAM
         totalRunnableThreads = 0;
     }
 
-    for (let row = 1; row < serverData.length; row++) {
+    for (let row = 1; row < serverData.length - 1; row++) {
         if (serverData[row][indexPurchasedByPlayer] === true) {
             totalRunnableThreads += Math.floor((serverData[row][indexMaxRam] - 0) / script_ram); // calculate available RAM on purchased servers
         }
@@ -151,35 +187,34 @@ function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack
 
     // ------------------ calculate server value ------------------
     let toHackServersTotalValue = 0; // total value of all the servers that we want to hack
-    for (let row = 1; row < serverData.length; row++) {
+    for (let row = 1; row < serverData.length - 1; row++) {
         toHackServersTotalValue += serverData[row][indexServerValue];
     }
 
     // ------------------ add server value as percentage to array ------------------
     serverData[0].push('percentageValue'); // add a new column to the header
-    for (let row = 1; row < serverData.length; row++) {
+    for (let row = 1; row < serverData.length - 1; row++) {
         serverData[row].push(serverData[row][indexServerValue] / toHackServersTotalValue);
     }
     let indexPercentageValue = serverData[0].indexOf('percentageValue'); // get the index of the new column
 
-    // ------------------ calculate threads for each server ------------------
-    let toHackServers = serverData.slice(1).filter(server => server[indexHackable] === true); // filter out servers that are not hackable
-    toHackServers.unshift(serverData[0]); // add the header back again so we can index it
-    toHackServers[0].push('threads'); // add a new column to the header
-    let indexThreads = toHackServers[0].indexOf('threads'); // get the index of the new column
-    for (let row = 1; row < toHackServers.length; row++) {
-        toHackServers[row].push(Math.floor(totalRunnableThreads * toHackServers[row][indexPercentageValue]));
+    // ------------------ calculate threads for each hackable server ------------------
+    serverData[0].push('threads'); // add a new column to the header
+    let indexThreads = serverData[0].indexOf('threads'); // get the index of the new column
+    for (let row = 1; row < serverData.length - 1; row++) {
+        serverData[row].push(Math.floor(totalRunnableThreads * serverData[row][indexPercentageValue]));
     }
 
-    // ------------------ kill running scripts ------------------
-    for (let row = 1; row < serverData.length; row++) {
-        if (serverData[row][indexPurchasedByPlayer] === true || serverData[row][indexHostname] === 'home') {
+    // ------------------ calculate threads for each purchased server ------------------
+    // todo
+
+    // ------------------ kill running scripts on purchased servers ------------------
+    for (let row = 1; row < serverData.length - 1; row++) {
+        if (serverData[row][indexPurchasedByPlayer] === true && serverData[row][indexHostname] !== 'home') {
             let processes = ns.ps(serverData[row][indexHostname]);
             for (let process of processes) {
                 if (process.filename.includes('hack')) { // kill all hacking processes
                     ns.kill(process.pid);
-                }
-                if (serverData[row][indexHostname] !== 'home') { // remove the hack script from all servers except 'home' anc copy a new version of it to the server
                     ns.rm(hackFile, serverData[row][indexHostname]);
                     ns.scp(hackFile, serverData[row][indexHostname]);
                 }
@@ -187,13 +222,34 @@ function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack
         }
     }
 
+    // ------------------ kill running scripts on home ------------------
+    for (let row = 1; row < serverData.length - 1; row++) {
+        if (serverData[row][indexHostname] === 'home' && hackOnHome) {
+            let processes = ns.ps(serverData[row][indexHostname]);
+            for (let process of processes) {
+                if (process.filename.includes('hack')) { // kill all hacking processes
+                    ns.kill(process.pid);
+                }
+            }
+            break;
+        }
+    }
+
     // ------------------ run hack scripts ------------------
     let remainingThreads = totalRunnableThreads;
+    let toHackServers = serverData.slice(1).filter(server => server[indexHackable] === true); // filter out servers that are not hackable
     let iH = 1; // Index to the current server we're executing hacking scripts on
+    toHackServers.unshift(serverData[0]); // we add the header to the toHackServers array
     let purchasedServers = serverData.slice(1).filter(server => server[indexPurchasedByPlayer] === true); // we build a new array with only our purchased servers
-    purchasedServers.unshift(serverData[0]); // we add the header to the purchasedServers array
     let iP = 1; // Index to the current purchased server we're executing hacking scripts on
+    purchasedServers.unshift(serverData[0]); // we add the header to the purchasedServers array
+    // if hackOnHome is false, we remove the 'home' server from the array
+    if (!hackOnHome) {
+        purchasedServers = purchasedServers.filter(server => server[indexHostname] !== 'home');
+    }
 
+    // ##############################################
+    // todo: debug this while loop, then we should be done!
     let counter = 1; // counter to prevent infinite loops, just in case we mess up something
     while (remainingThreads > 0) {
         if (iP >= purchasedServers.length || iH >= toHackServers.length) {
@@ -203,7 +259,11 @@ function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack
         }
 
         // if the current purchased server has enough threads available, we run the hack script on it with the given amount of threads
+        ns.tprint("\n\n");
+        ns.tprint(`Purchased server: ${purchasedServers[iP][indexHostname]}, toHack server: ${toHackServers[iH][indexHostname]}`);
+        ns.tprint(`Purchased server threads: ${purchasedServers[iP][indexThreads]}, toHack server threads: ${toHackServers[iH][indexThreads]}`);
         if (purchasedServers[iP][indexThreads] > toHackServers[iH][indexThreads]) {
+            ns.tprint(" purchased server > to hack server");
             if (toHackServers[iH][indexThreads] > 0) { // only if threads are > 0 we run the hack script
                 ns.exec(hackFile, purchasedServers[iP][indexHostname], toHackServers[iH][indexThreads], toHackServers[iH][indexHostname], toHackServers[iH][indexThreads]);
                 purchasedServers[iP][indexThreads] -= toHackServers[iH][indexThreads];
@@ -211,6 +271,7 @@ function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack
             }
             iH++; // go to the next toHackServer
         } else { // if not, we run the hack script with as many threads as possible and the next server can use the remaining threads
+            ns.tprint(" purchased server < to hack server");
             if (purchasedServers[iP][indexThreads] > 0) { // only if threads are > 0 we run the hack script
                 ns.exec(hackFile, purchasedServers[iP][indexHostname], purchasedServers[iP][indexThreads], toHackServers[iH][indexHostname], purchasedServers[iP][indexThreads]);
                 toHackServers[iH][indexThreads] -= purchasedServers[iP][indexThreads];
@@ -222,21 +283,6 @@ function hackOnPurchasedServers(ns, freeRAMonHome, hackFile, scanFile) { // hack
         if (counter > 1000 || iP >= purchasedServers.length || iH >= toHackServers.length) {
             ns.tprint("Error 2053: Counter exceeded 1000 or index exceeded array length. Breaking loop.");
             break;
-        }
-    }
-}
-
-function serverHackable(target, serverData) { // check if the target server is hackable
-    const index_hostname = serverData[0].indexOf('hostname');
-    const index_admin = serverData[0].indexOf('hasAdminRights');
-
-    for (let i = 1; i < serverData.length - 1; i++) {
-        if (serverData[i][index_hostname] === target) {
-            if (serverData[i][index_admin]) {
-                return true;
-            } else {
-                return false;
-            }
         }
     }
 }
